refactor(perf): separate capture and trace summaries
